Fixed the function and made new struct for char and buffer
